fix: InvalidCentralSystemCertificate was logged on every failed connection (#1156)
A CSMS certificate may be considered invalid when a charger has the incorrect
time. This can be recovered by using NTP. However a security event is raised
for each failed connection attempt until the correct time is set.

This change ensures that InvalidCentralSystemCertificate is logged only
on the first occurance.

Signed-off-by: James Chapman <james.chapman@pionix.de>